feat: Audit and enhance Cypher query engine
Fix 5 critical bugs:
- Fix operator determination in comparison/arithmetic expressions by
  walking parse tree children in order instead of cumulative token counts
- Preserve integer precision in arithmetic (int64_t ops stay int64_t)
- Implement three-valued NULL logic for AND/OR/XOR operators
- Support string concatenation via + operator
- Fix unary minus to preserve integer type

Code quality improvements:
- Remove dead code in buildMatchPattern, properly handle CartesianProduct
- Remove deprecated buildWherePredicate API, migrate to buildExpression
- Fix fragile merge handler with type-safe CypherLexer token matching
- Remove legacy string-based ProjectItem constructor and parseExpression

New features:
- Add STARTS WITH, ENDS WITH, CONTAINS string comparison operators
  with grammar update and ANTLR parser regeneration
- Implement expression-based UNWIND (e.g. UNWIND n.ids AS x)
- Add MergeEdgeOperator for MERGE edge pattern support

+++ tests/src/query/expressions/test_Expressions_Coverage.cpp
@@ -628,7 +628,7 @@ TEST_F(ExpressionsCoverageTest, UnaryOpExpression_MINUS_Variants) {
 		UnaryOpExpression expr(UnaryOperatorType::UOP_MINUS, std::move(operand));
 		ExpressionEvaluator evaluator(*context_);
 		expr.accept(evaluator);
-		EXPECT_EQ(std::get<double>(evaluator.getResult().getVariant()), -42.0);
+		EXPECT_EQ(std::get<int64_t>(evaluator.getResult().getVariant()), -42);
 	}
 
 	{
@@ -1373,9 +1373,9 @@ TEST_F(ExpressionsCoverageTest, ListComprehensionExpression_CompleteCoverage) {
 		expr.accept(evaluator);
 		const auto& result = evaluator.getResult().getList();
 		EXPECT_EQ(result.size(), 3ULL);
-		EXPECT_EQ(std::get<double>(result[0].getVariant()), 2.0);
-		EXPECT_EQ(std::get<double>(result[1].getVariant()), 4.0);
-		EXPECT_EQ(std::get<double>(result[2].getVariant()), 6.0);
+		EXPECT_EQ(std::get<int64_t>(result[0].getVariant()), 2);
+		EXPECT_EQ(std::get<int64_t>(result[1].getVariant()), 4);
+		EXPECT_EQ(std::get<int64_t>(result[2].getVariant()), 6);
 	}
 
 	// EXTRACT with WHERE and MAP: [x IN list WHERE x > 2 | x * 3]
@@ -1394,9 +1394,9 @@ TEST_F(ExpressionsCoverageTest, ListComprehensionExpression_CompleteCoverage) {
 		expr.accept(evaluator);
 		const auto& result = evaluator.getResult().getList();
 		EXPECT_EQ(result.size(), 3ULL); // 3, 4, 5
-		EXPECT_EQ(std::get<double>(result[0].getVariant()), 9.0);
-		EXPECT_EQ(std::get<double>(result[1].getVariant()), 12.0);
-		EXPECT_EQ(std::get<double>(result[2].getVariant()), 15.0);
+		EXPECT_EQ(std::get<int64_t>(result[0].getVariant()), 9);
+		EXPECT_EQ(std::get<int64_t>(result[1].getVariant()), 12);
+		EXPECT_EQ(std::get<int64_t>(result[2].getVariant()), 15);
 	}
 
 	// REDUCE - Currently returns list like FILTER (REDUCE not fully implemented)
